requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary
paho-mqtt==1.6.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
python-dotenv==0.20.0
